Add jest-circus `teardown` as a failsafe.
Anecdotally, I've seen some weird interactions when running more than
one Polly-powered tests in a single runner (but maybe multi-threaded?).

This seemed to fix it, though I'm not 100% sure what the underlying
cause is.

In either case, this still seems like a reasonable thing to have in
place, just as a fallback.